feat(api): take into account size_hierarchy
In grouped_geom_by_areas mode to filter out observations which blurring
geometry size is bigger than the aggregation area provided.

Example:
Aggregation area = M1 : all blurred obs with a geometry which size is
bigger than the size of M1 are not selected. Nevertheless, they are
still displayed in the list but when selected them, a message in
frontend appears saying that the user must go back in point mode to view
this obs